requirements.txt
Django>=5.1,<6.0
djangorestframework>=3.15,<4.0
drf-spectacular>=0.28,<1.0
django-environ>=0.11,<1.0
django-cors-headers>=4.4,<5.0
psycopg[binary]>=3.2,<4.0
pgvector>=0.3,<1.0
redis>=5.2,<6.0
django-redis>=5.4,<6.0
celery>=5.4,<6.0
python-json-logger>=2.0,<3.0
gunicorn>=22.0,<23.0
uvicorn>=0.30,<1.0
sentence-transformers>=3.0,<4.0
